move initialization of network data into the cpp
