+++ networks/iso7/Make.package
@@ -2,8 +2,7 @@ F90EXE_sources += actual_network.F90
 F90EXE_sources += network_properties.F90
 CEXE_headers += network_properties.H
 
-CEXE_sources += actual_network.cpp
-CEXE_headers += actual_network_data.H
+CEXE_sources += actual_network_data.cpp
 CEXE_headers += actual_network.H
 
 ifeq ($(USE_REACT),TRUE)
@@ -12,7 +11,7 @@ F90EXE_sources += actual_burner.F90
 endif
 F90EXE_sources += actual_rhs.F90
 
-CEXE_sources += actual_rhs.cpp
+CEXE_sources += actual_rhs_data.cpp
 CEXE_headers += actual_rhs.H
 
 USE_RATES       = TRUE
